fix(codeReview): code review fixes
